refactor : 지하철 역에 대한 인수 테스트 시나리오에서 행위에 집중할 수 있도록 반환 정보와 Step Definitions 메서드명을 수정한다
