+++ src/test/java/subway/station/StationTestStepDefinition.java
@@ -1,0 +1,63 @@
+package subway.station;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class StationTestStepDefinition {
+    public static StationResponse 지하철_역_생성_요청(String name) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", name);
+
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+            .body(params)
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .when().post("/stations")
+            .then().log().all()
+            .extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+
+        return response.as(StationResponse.class);
+    }
+
+    public static StationResponse 지하철_역_생성_요청_역_정보_반환(String name) {
+        return 지하철_역_생성_요청(name);
+    }
+
+    public static List<StationResponse> 지하철_역_목록_조회_요청() {
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .when().get("/stations")
+            .then().log().all()
+            .extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+
+        return response.jsonPath().getList(".", StationResponse.class);
+    }
+
+    public static List<String> 지하철_역_이름_목록_조회_요청() {
+        return 지하철_역_목록_조회_요청().stream()
+            .map(StationResponse::getName)
+            .collect(java.util.stream.Collectors.toList());
+    }
+
+    public static void 지하철_역_삭제_요청(Long id) {
+        ExtractableResponse<Response> response = RestAssured.given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .when().delete("/stations/" + id)
+            .then().log().all()
+            .extract();
+
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+    }
+}
